+++ CHANGELOG.md
@@ -9,12 +9,15 @@ This document outlines major changes between releases.
 
 ### Changed
 - Default polling interval for container operations from fixed 1s to dynamic value based on block time (#1194)
+- Use single node connection instead of pool to create bucket (NeoFS) and set EACL (NeoFS) for create bucket operation (#1198)
 
 ### Fixed
 - Incorrect EpochDuration processing wrt API 2.18 leading to improper expiration calculations (#1193)
+- Incorrect bucket owner usage for search operations (#1184)
 
 ### Updated
 - NeoFS SDK to RC14 version (#1202)
+- github.com/nspcc-dev/neo-go v0.110.0 => v0.111.0 (#1184)
 
 ### Removed